Part 3: Generating a dungeon
Added a dungeon generator that is almost identical to the Python
tutorial. Quite a lot of JavaScript-specific changes to the code
needed—some nicer than Python, some not so nice. It works! It produces
nice empty little layouts with boxes of varying sizes and straight line
passageways cutting through and connecting them.

Also added `rung` as a dependency to get a decent integer range API for
random number generation.