Part 4: Field of View
Finally got around to adding FOV, based on a JavaScript port of SquidLib
with some restructuring to ensure there aren’t any dependencies on the
Stage or game state and no OO code required.

Most of this tutorial focuses on adding FOV as a game mechanic, rather
than how the algorithm is implemented technically. It ends up being
pretty straightforward with JavaScript sets and cleanly drops in over
the top of the existing codebase with no major modifications needed.