Part 5: Placing monsters and bumping them around
Still following the structure of the existing tutorial but changing the
implementation quite a bit in some places. Part 5 sets up the foundation
for placing entities on the map, holding them in a spatial lookup table
and adding tile occupancy checks that will form the basis for bump
combat.

I was going to add some action handling code here too, but ended up
backing out of it and reverting back to what the tutorial does with an
update loop that bounces control between player and entities. This will
probably be deleted very soon, but I felt that adding any more
complexity and detail to this particular tutorial step would have been a
bit too much. We’ll tackle this in the next one.