Part 6: Dealing damage and deciding direction
Epic update this time round. Had very little time to work on this, so
have fallen behind on progress a bit, but have gotten most of the big
annoying bits of scaffolding and setup out of the way now.

This part adds a simple combat system and also incorporates pathfinding.
Might be a little too much for one update, but it seems to fit together
quite well thematically.

This also marks the first real introduction of game mechanics. There are
rules for monster AI, rules for collisions between entities moving
across tiles and rules determining what constitutes an attack.

Not much to go on as yet, but there’s huge opportunity to take this as a
base and rework it into something exciting and fun. In the meantime, I
need to think about some actual games to make on this foundation. One of
the downsides of the existing tutorial is that it encourages leaping
ahead without planning or game design thinking. I’ve tried to spend a
bit more time exploring design goals and suggesting alternative
strategies for implementation, but it would be really good to work on a
bit of genuine content rather than filler.